+++ pyswtools/stat/main.py
@@ -11,18 +11,18 @@ from ..utils import check_system, check_system_verbose
 from ..config import get_config
 
 
-class TypeOutput(Enum):
+class TypeOutput(str, Enum):
     """Class represeting an output type"""
 
-    TREE = 1
-    LIST = 2
+    TREE = "tree"
+    LIST = "list"
 
 
-class TypeSort(Enum):
+class TypeSort(str, Enum):
     """Class represeting an output type"""
 
-    MASS = 1
-    NAME = 2
+    MASS = "mass"
+    NAME = "name"
 
 
 if check_system():
@@ -39,9 +39,9 @@ def sort_key_struct(struct: dict, type_sort: TypeSort = TypeSort.MASS) -> list:
     """
     Sort a dict struct following a given sort
     """
-    if type_sort == TypeSort.NAME:
+    if type_sort is TypeSort.NAME:
         return sorted(struct.keys(), key=str.lower)
-    if type_sort == TypeSort.MASS:
+    if type_sort is TypeSort.MASS:
         return sorted(
             struct.keys(),
             key=lambda i: struct[i]["mass"] * struct[i]["number"],
@@ -51,27 +51,44 @@ def sort_key_struct(struct: dict, type_sort: TypeSort = TypeSort.MASS) -> list:
     return struct.keys()
 
 
+def print_header():
+    header = f"{'Name':<50} | {'Mtot':<5} | {'n':<3} | {'Mpart':<6} | {'Density':<7}  "
+    click.echo(header)
+    click.echo("-" * len(header))
+
+
 def display_tree(
-    tree_struct_mass: dict, type_sort: TypeSort = TypeSort.NAME, indent: int = 0
+    tree_struct_mass: dict, type_sort: TypeSort = TypeSort.NAME, indent: str = ""
 ) -> None:
     """
     Display a tree struct recursively
     """
-    for elem in sort_key_struct(tree_struct_mass, type_sort):
+    sorted_list = sort_key_struct(tree_struct_mass, type_sort)
+    n = len(sorted_list)
+    if indent == "":
+        print_header()
+    for idx, elem in enumerate(sorted_list):
+        char = "┝━"
+        if idx == n - 1:
+            char = "┕━"
         click.echo(
-            f"{'  ' * indent + '- ' + elem :<50} : {tree_struct_mass[elem]['mass'] * tree_struct_mass[elem]['number']:5.3f} ({tree_struct_mass[elem]['number']:3d} x {tree_struct_mass[elem]['mass']:6.4f})"
+            f"{indent + char + ' ' + elem :<50} | {tree_struct_mass[elem]['mass'] * tree_struct_mass[elem]['number']:5.3f} | {tree_struct_mass[elem]['number']:3d} | {tree_struct_mass[elem]['mass']:6.4f} | {tree_struct_mass[elem]['density']:6.4f}"
         )
         if len(tree_struct_mass[elem]["children"]) > 0:
-            display_tree(tree_struct_mass[elem]["children"], type_sort, indent + 1)
+            char = "│ "
+            if idx == n - 1:
+                char = "  "
+            display_tree(tree_struct_mass[elem]["children"], type_sort, indent + char)
 
 
 def display_list(list_struct: dict, type_sort: TypeSort = TypeSort.NAME) -> None:
     """
     Display a list struct
     """
+    print_header()
     for elem in sort_key_struct(list_struct, type_sort):
         click.echo(
-            f"{'- ' + elem :<50} : {list_struct[elem]['mass'] * list_struct[elem]['number']:5.3f} ({list_struct[elem]['number']:3d} x {list_struct[elem]['mass']:6.4f})"
+            f"{'- ' + elem :<50} | {list_struct[elem]['mass'] * list_struct[elem]['number']:5.3f} | {list_struct[elem]['number']:3d} | {list_struct[elem]['mass']:6.4f} | {list_struct[elem]['density']:6.4f}"
         )
 
 
@@ -120,7 +137,11 @@ def complete_info_assembly(sw_comp, dict_of_comp: dict) -> dict:
         sw_mass = sw_comp_doc_ext.CreateMassProperty2.Mass
 
         # Create an new entity in the general dict
-        dict_of_comp[sw_comp_name] = {"number": 1, "mass": sw_mass}
+        dict_of_comp[sw_comp_name] = {
+            "number": 1,
+            "mass": sw_mass,
+            "density": sw_comp_doc_ext.CreateMassProperty2.density,
+        }
 
     # Get info about children
     sw_comp_children = sw_comp.GetChildren
@@ -129,6 +150,7 @@ def complete_info_assembly(sw_comp, dict_of_comp: dict) -> dict:
     # Create the entry in the tree dict
     return {
         "mass": dict_of_comp[sw_comp_name]["mass"],
+        "density": dict_of_comp[sw_comp_name]["density"],
         "number": 1,
         "children": children,
     }
@@ -143,10 +165,15 @@ def complete_info_assembly(sw_comp, dict_of_comp: dict) -> dict:
         dir_okay=False,
     ),
 )
-@click.option("--tree", "type_output", flag_value=TypeOutput.TREE, default=True)
-@click.option("--list", "type_output", flag_value=TypeOutput.LIST)
-@click.option("--mass", "type_sort", flag_value=TypeSort.MASS, default=True)
-@click.option("--name", "type_sort", flag_value=TypeSort.NAME)
+@click.option(
+    "--type_output",
+    "type_output",
+    type=click.Choice(TypeOutput),
+    default=TypeOutput.TREE,
+)
+@click.option(
+    "--type_sort", "type_sort", type=click.Choice(TypeSort), default=TypeSort.MASS
+)
 def stat(input_path: str, type_output: TypeOutput, type_sort: TypeSort) -> None:
     """
     Display stat about an assembly
@@ -177,14 +204,29 @@ def stat(input_path: str, type_output: TypeOutput, type_sort: TypeSort) -> None:
     sw_comps = sw_doc.GetComponents(True)
 
     # Empty struct to fill
-    dict_of_comp = {}
+    assembly_name = filename.rpartition(".")[-0]
+    dict_of_comp = {
+        f"{assembly_name}": {
+            "mass": sw_doc.Extension.CreateMassProperty2.Mass,
+            "density": sw_doc.Extension.CreateMassProperty2.Density,
+            "number": 1,
+        }
+    }
 
-    # Get all the infos
-    tree_of_infos = complete_info_on_list(sw_comps, dict_of_comp)
+    tree_of_comp = {
+        f"{assembly_name}": {
+            "mass": dict_of_comp[assembly_name]["mass"],
+            "density": dict_of_comp[assembly_name]["density"],
+            "number": 1,
+            "children": complete_info_on_list(sw_comps, dict_of_comp),
+        }
+    }
 
-    if type_output == TypeOutput.TREE:
-        display_tree(tree_of_infos, type_sort)
-    elif type_output == TypeOutput.LIST:
+    if type_output is TypeOutput.TREE:
+        display_tree(tree_of_comp, type_sort)
+    elif type_output is TypeOutput.LIST:
         display_list(dict_of_comp, type_sort)
     else:
-        pass
+        click.echo(
+            f"Type output {type_output} is unknown {type_output is TypeOutput.TREE} {type(type_output)}"
+        )